backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
python-jose
passlib[bcrypt]
python-dotenv
pytest
httpx
pytest-asyncio
databases
